backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0